design-audit: move ObservationGridCard skeleton to common/, dedupe CardContent sx
ProfileObservationCardSkeleton lived under profile/ but was really the
loading placeholder for common/ObservationGridCard, and feed/FeedView.tsx
was importing it across the profile/ domain boundary for ExploreGridCard.
Renamed to common/ObservationGridCardSkeleton and exported the CardContent
sx literal from ObservationGridCard so the skeleton and real card can't
drift apart.
