+++ frontend/src/components/common/ObservationGridCardSkeleton.tsx
@@ -1,9 +1,10 @@
 import { Box, Card, CardContent, Skeleton, Typography } from "@mui/material";
+import { observationGridCardContentSx } from "./ObservationGridCard";
 
 /**
- * Skeleton for profile observation card grid items
+ * Loading placeholder for ObservationGridCard.
  */
-export function ProfileObservationCardSkeleton() {
+export function ObservationGridCardSkeleton() {
   return (
     <Card sx={{ display: "flex", flexDirection: "column" }}>
       <Box sx={{ position: "relative", aspectRatio: "1", width: "100%" }}>
@@ -12,7 +13,7 @@ export function ProfileObservationCardSkeleton() {
           sx={{ position: "absolute", inset: 0, width: "100%", height: "100%" }}
         />
       </Box>
-      <CardContent sx={{ p: 1.5, "&:last-child": { pb: 1.5 }, flex: 1 }}>
+      <CardContent sx={observationGridCardContentSx}>
         <Typography variant="body2">
           <Skeleton width="70%" />
         </Typography>
